fix: add redis in methods

+++ app/modules/favorites/graphql.py
@@ -1,7 +1,6 @@
 from uuid import UUID
 import strawberry
 from app.modules.product.graphql import ProductType
-from app.core.dependencies import get_current_user_graphql
 from app.modules.favorites.service import FavoriteService
 from app.modules.favorites.schemas import FavoritesCreate
 
@@ -18,7 +17,10 @@ class FavoritesQuery:
     @strawberry.field
     async def favorites(self, info: strawberry.Info, limit: int = 20, offset: int = 0) -> list[FavoriteType]:
         db = info.context["db"]
-        current_user = await get_current_user_graphql(info, db)
+        current_user = info.context["user"]
+        if not current_user:
+            raise ValueError("401: Could not validate credentials")
+            
         redis_client = info.context["redis"]
         fav_lists = await FavoriteService.get_favorites_by_user(redis_client, db, current_user.id, limit, offset)
         return [
@@ -42,7 +44,10 @@ class FavoritesQuery:
     @strawberry.field
     async def get_favorite_list_by_id(self, info: strawberry.Info, id: UUID) -> FavoriteType | None:
         db = info.context["db"]
-        current_user = await get_current_user_graphql(info, db)
+        current_user = info.context["user"]
+        if not current_user:
+            raise ValueError("401: Could not validate credentials")
+            
         fav_list = await FavoriteService.get_favorite_list_by_id(db, id, current_user.id)
         if not fav_list:
             return None
@@ -72,10 +77,13 @@ class FavoritesMutation:
         description: str | None = None
     ) -> FavoriteType:
         db = info.context["db"]
-        current_user = await get_current_user_graphql(info, db)
-        
+        current_user = info.context["user"]
+        redis_client = info.context["redis"]
+        if not current_user:
+            raise ValueError("401: Could not validate credentials")
+            
         fav_data = FavoritesCreate(name=name, description=description)
-        new_list = await FavoriteService.create_favorite_list(db, current_user.id, fav_data)
+        new_list = await FavoriteService.create_favorite_list(db, redis_client, current_user.id, fav_data)
         
         return FavoriteType(
             id=new_list.id,
@@ -88,7 +96,10 @@ class FavoritesMutation:
     @strawberry.mutation
     async def delete_favorite_list(self, info: strawberry.Info, id: UUID) -> bool:
         db = info.context["db"]
-        current_user = await get_current_user_graphql(info, db)
+        current_user = info.context["user"]
+        if not current_user:
+            raise ValueError("401: Could not validate credentials")
+            
         return await FavoriteService.delete_favorite_list(db, id, current_user.id)
 
     @strawberry.mutation
@@ -99,8 +110,10 @@ class FavoritesMutation:
         product_id: UUID
     ) -> FavoriteType:
         db = info.context["db"]
-        current_user = await get_current_user_graphql(info, db)
-        
+        current_user = info.context["user"]
+        if not current_user:
+            raise ValueError("401: Could not validate credentials")
+            
         updated_list = await FavoriteService.add_product_to_favorites(
             db, favorite_id, current_user.id, product_id
         )
@@ -131,8 +144,10 @@ class FavoritesMutation:
         product_id: UUID
     ) -> FavoriteType:
         db = info.context["db"]
-        current_user = await get_current_user_graphql(info, db)
-        
+        current_user = info.context["user"]
+        if not current_user:
+            raise ValueError("401: Could not validate credentials")
+            
         updated_list = await FavoriteService.remove_product_from_favorites(
             db, favorite_id, current_user.id, product_id
         )
